Added Gemfile.
I haven't done anything with this project for a while and coming back to it was a bit of a pain because I had to get gems installed through trial and error.

I've specified the latest versions of the gems that were available at the time of my last post (21st Apr 2011).

I've run `bundle exec webby` and compared the output to the copy of my site downloaded from seagul.co.uk.  They're close enough to give me confidence that I'm locking to the correct gem versions.